feat(APIM-576): relax buyers name to match exporter name rules
