+++ src/constants/buyer-name.constant.ts
@@ -1,3 +1,5 @@
+import { SHAREPOINT } from './sharepoint.constant';
+
 export const BUYER_NAME = {
-  REGEX: /^(?!\s)[\w\-.()\s]+(?<![\s.])$/,
+  REGEX: SHAREPOINT.SAFE_FOLDER_NAME.REGEX,
 };
