+++ test/site-buyer/post-site-buyers.api-test.ts
@@ -1,7 +1,7 @@
 import { IncorrectAuthArg, withClientAuthenticationTests } from '@ukef-test/common-tests/client-authentication-api-tests';
 import { withCustodianCreateAndProvisionErrorCasesApiTests } from '@ukef-test/common-tests/custodian-create-and-provision-error-cases-api-tests';
+import { withBuyerNameFieldValidationApiTests } from '@ukef-test/common-tests/request-field-validation-api-tests/buyer-name-field-validation-api-tests';
 import { withExporterNameFieldValidationApiTests } from '@ukef-test/common-tests/request-field-validation-api-tests/exporter-name-field-validation-api-tests';
-import { withSharepointResourceNameFieldValidationApiTests } from '@ukef-test/common-tests/request-field-validation-api-tests/sharepoint-resource-name-field-validation-api-tests';
 import { withSiteIdParamValidationApiTests } from '@ukef-test/common-tests/request-param-validation-api-tests/site-id-param-validation-api-tests';
 import { withSharedGraphExceptionHandlingTests } from '@ukef-test/common-tests/shared-graph-exception-handling-api-tests';
 import { Api } from '@ukef-test/support/api';
@@ -218,8 +218,7 @@ describe('POST /sites/{siteId}/buyers', () => {
   });
 
   describe('field validation', () => {
-    withSharepointResourceNameFieldValidationApiTests({
-      fieldName: 'buyerName',
+    withBuyerNameFieldValidationApiTests({
       valueGenerator,
       validRequestBody: createBuyerFolderRequest,
       makeRequest: (body: unknown[]) => makeRequestWithBody(body),
