+++ test/site-deal/post-site-deal-facilities.api-test.ts
@@ -2,9 +2,9 @@ import { UKEFID } from '@ukef/constants';
 import { UkefId } from '@ukef/helpers';
 import { IncorrectAuthArg, withClientAuthenticationTests } from '@ukef-test/common-tests/client-authentication-api-tests';
 import { withCustodianCreateAndProvisionErrorCasesApiTests } from '@ukef-test/common-tests/custodian-create-and-provision-error-cases-api-tests';
+import { withBuyerNameFieldValidationApiTests } from '@ukef-test/common-tests/request-field-validation-api-tests/buyer-name-field-validation-api-tests';
 import { withExporterNameFieldValidationApiTests } from '@ukef-test/common-tests/request-field-validation-api-tests/exporter-name-field-validation-api-tests';
 import { withFacilityIdentifierFieldValidationApiTests } from '@ukef-test/common-tests/request-field-validation-api-tests/facility-identifier-validation-api-tests';
-import { withSharepointResourceNameFieldValidationApiTests } from '@ukef-test/common-tests/request-field-validation-api-tests/sharepoint-resource-name-field-validation-api-tests';
 import { withSiteIdParamValidationApiTests } from '@ukef-test/common-tests/request-param-validation-api-tests/site-id-param-validation-api-tests';
 import { withParamValidationApiTests } from '@ukef-test/common-tests/request-param-validation-api-tests/string-param-validation-api-tests';
 import { withSharedGraphExceptionHandlingTests } from '@ukef-test/common-tests/shared-graph-exception-handling-api-tests';
@@ -300,8 +300,7 @@ describe('Create Site Deal Facility Folder', () => {
       givenAnyRequestBodyWouldSucceed: () => givenAnyRequestBodyWouldSucceed(),
     });
 
-    withSharepointResourceNameFieldValidationApiTests({
-      fieldName: 'buyerName',
+    withBuyerNameFieldValidationApiTests({
       valueGenerator,
       validRequestBody: createFacilityFolderRequestDto,
       successStatusCode: 201,
